+++ stream-chat-android-ui-common/src/main/kotlin/com/getstream/sdk/chat/CaptureMediaContract.kt
@@ -58,7 +58,7 @@ public class CaptureMediaContract : ActivityResultContract<Unit, File?>() {
             }
         val intents = takePictureIntents + recordVideoIntents
         val initialIntent = intents.lastOrNull() ?: Intent()
-        return Intent.createChooser(initialIntent, context.getString(R.string.stream_ui_message_input_capture_media))
+        return Intent.createChooser(initialIntent, context.getString(R.string.stream_ui_message_composer_capture_media))
             .apply {
                 putExtra(
                     Intent.EXTRA_INITIAL_INTENTS,