+++ stream-chat-android-ui-components-sample/src/main/kotlin/io/getstream/chat/ui/sample/feature/channel/add/AddChannelFragment.kt
@@ -23,10 +23,10 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
-import com.getstream.sdk.chat.viewmodel.MessageInputViewModel
 import com.getstream.sdk.chat.viewmodel.messages.MessageListViewModel
 import io.getstream.chat.android.livedata.utils.EventObserver
-import io.getstream.chat.android.ui.message.input.viewmodel.bindView
+import io.getstream.chat.android.ui.message.composer.viewmodel.MessageComposerViewModel
+import io.getstream.chat.android.ui.message.composer.viewmodel.bindView
 import io.getstream.chat.android.ui.message.list.viewmodel.bindView
 import io.getstream.chat.android.ui.message.list.viewmodel.factory.MessageListViewModelFactory
 import io.getstream.chat.ui.sample.R
@@ -70,10 +70,15 @@ class AddChannelFragment : Fragment() {
     private fun initializeChannel(cid: String) {
         val factory = MessageListViewModelFactory(cid)
         val messageListViewModel = factory.create(MessageListViewModel::class.java)
-        val messageInputViewModel = factory.create(MessageInputViewModel::class.java)
+        val messageComposerViewModel = factory.create(MessageComposerViewModel::class.java)
         binding.addChannelView.apply {
             messageListViewModel.bindView(messageListView, viewLifecycleOwner)
-            messageInputViewModel.bindView(messageInputView, viewLifecycleOwner)
+            messageComposerViewModel.bindView(messageComposerView, viewLifecycleOwner)
+
+            messageComposerView.sendMessageButtonClickListener = {
+                messageComposerViewModel.sendMessage(messageComposerViewModel.buildNewMessage())
+                addChannelViewModel.onEvent(AddChannelViewModel.Event.MessageSent)
+            }
         }
     }
 
@@ -89,7 +94,8 @@ class AddChannelFragment : Fragment() {
                     )
                     AddChannelViewModel.State.Loading,
                     is AddChannelViewModel.State.Result,
-                    is AddChannelViewModel.State.ResultMoreUsers -> Unit
+                    is AddChannelViewModel.State.ResultMoreUsers,
+                    -> Unit
                 }
             }
             errorEvents.observe(
@@ -107,9 +113,6 @@ class AddChannelFragment : Fragment() {
             }
             setOnCreateGroupButtonListener {
                 findNavController().navigateSafely(R.id.action_addChannelFragment_to_addGroupChannelFragment)
-            }
-            messageInputView.setOnSendButtonClickListener {
-                addChannelViewModel.onEvent(AddChannelViewModel.Event.MessageSent)
             }
         }
     }