+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/composer/content/DefaultMessageComposerMentionSuggestionsContent.kt
@@ -25,7 +25,6 @@ import androidx.core.view.isVisible
 import androidx.recyclerview.widget.RecyclerView
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.common.composer.MessageComposerState
-import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.ui.common.extensions.internal.createStreamThemeWrapper
 import io.getstream.chat.android.ui.common.extensions.internal.streamThemeInflater
 import io.getstream.chat.android.ui.common.internal.SimpleListAdapter
@@ -40,7 +39,6 @@ import io.getstream.chat.android.ui.utils.extensions.applyTint
 /**
  * Represents the default mention suggestion list popup shown above [MessageComposerView].
  */
-@ExperimentalStreamChatApi
 public class DefaultMessageComposerMentionSuggestionsContent : FrameLayout, MessageComposerContent {
     /**
      * Generated binding class for the XML layout.