+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/adapter/viewholder/internal/GiphyViewHolder.kt
@@ -26,6 +26,7 @@ import io.getstream.chat.android.common.message.list.CancelGiphy
 import io.getstream.chat.android.common.message.list.SendGiphy
 import io.getstream.chat.android.common.message.list.ShuffleGiphy
 import io.getstream.chat.android.ui.common.extensions.internal.streamThemeInflater
+import io.getstream.chat.android.ui.common.style.setTextStyle
 import io.getstream.chat.android.ui.databinding.StreamUiItemMessageGiphyBinding
 import io.getstream.chat.android.ui.message.list.GiphyViewHolderStyle
 import io.getstream.chat.android.ui.message.list.adapter.MessageListItemPayloadDiff
@@ -103,11 +104,11 @@ internal class GiphyViewHolder(
 
             giphyIconImageView.setImageDrawable(style.giphyIcon)
 
-            style.labelTextStyle.apply(giphyLabelTextView)
-            style.queryTextStyle.apply(giphyQueryTextView)
-            style.cancelButtonTextStyle.apply(cancelButton)
-            style.shuffleButtonTextStyle.apply(shuffleButton)
-            style.sendButtonTextStyle.apply(sendButton)
+            giphyLabelTextView.setTextStyle(style.labelTextStyle)
+            giphyQueryTextView.setTextStyle(style.queryTextStyle)
+            cancelButton.setTextStyle(style.cancelButtonTextStyle)
+            shuffleButton.setTextStyle(style.shuffleButtonTextStyle)
+            sendButton.setTextStyle(style.sendButtonTextStyle)
         }
     }